+++ src/engine/ray_tracer.c
@@ -6,7 +6,7 @@
 /*   By: iherman- <iherman-@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/02 11:07:06 by iherman-          #+#    #+#             */
-/*   Updated: 2025/10/17 14:33:07 by iherman-         ###   ########.fr       */
+/*   Updated: 2025/10/24 21:27:17 by iherman-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,6 @@ static unsigned int	intersect_test(t_ray ray, t_rt_data *data)
 		if (obj->check_intersection(obj->obj, ray, &hitinfo) && hitinfo.t < best_hit.t)
 		{
 			best_hit = hitinfo;
-			//printf("t= %f\n", best_hit.t);
 			hit_found = true;
 		}
 		obj = obj->next;
@@ -35,38 +34,25 @@ static unsigned int	intersect_test(t_ray ray, t_rt_data *data)
 	if (hit_found)
 	{
 		unsigned int color = calculate_color(&best_hit, data);
-		//printf("color: %u\n", color);
 		return (color);
 	}
 	return (0 << 24 | 0 << 16 | 0 << 8 | 255);
 }
 
-t_ray	generate_ray(t_rt_data *data, int screenX, int screenY)
+t_ray	generate_ray(t_rt_data *data, int screen_x, int screen_y)
 {
 	t_ray ray;
-
-	double aspectRatio = (double)WINDOW_WIDTH / (double)WINDOW_HEIGHT;
-	double ndcX = (2.0 * (screenX + 0.5) / WINDOW_WIDTH) - 1.0;
-	double ndcY = 1.0 - (2.0 * (screenY + 0.5) / WINDOW_HEIGHT);
-	double fovAdjust = tan(data->camera.fov * 0.5 * 3.14 / 180.0);
+	const double ndcX = (2.0 * (screen_x + 0.5) / WINDOW_WIDTH) - 1.0;
+	const double ndcY = 1.0 - (2.0 * (screen_y + 0.5) / WINDOW_HEIGHT);
 
 	t_vec3 rayDir = vector_add(
 		data->camera.norm_vec,
 		vector_add(
-			vector_multiply(data->camera.right, ndcX * aspectRatio * fovAdjust),
-			vector_multiply(data->camera.up, ndcY * fovAdjust)));
+			vector_multiply(data->camera.right, ndcX * data->camera.aspect_ratio * data->camera.fov_adjust),
+			vector_multiply(data->camera.up, ndcY * data->camera.fov_adjust)));
 	ray.origin = data->camera.viewpoint;
 	ray.direction = normalize(rayDir);
 	return (ray);
-}
-
-unsigned int	shoot_ray(t_rt_data *data, int screen_x, int screen_y)
-{
-	t_ray			ray;
-
-	ray = generate_ray(data, screen_x, screen_y);
-	//printf("ray generated \n"); -> functional
-	return (intersect_test(ray, data));
 }
 
 void	raytracer(void *tmp_data)
@@ -75,6 +61,7 @@ void	raytracer(void *tmp_data)
 	unsigned int	color;
 	int				screen_x;
 	int				screen_y;
+	t_ray			ray;
 
 	data = (t_rt_data *)tmp_data;
 	screen_x = 0;
@@ -83,7 +70,8 @@ void	raytracer(void *tmp_data)
 	{
 		while (screen_y < WINDOW_HEIGHT)
 		{
-			color = shoot_ray(data, screen_x, screen_y);
+			ray = generate_ray(data, screen_x, screen_y);
+			color = intersect_test(ray, data);
 			mlx_put_pixel(data->mlx_img, screen_x, screen_y, color);
 			screen_y++;
 		}